Added screenshots demonstrating functionality

+++ core/views.py
@@ -129,7 +129,9 @@ def delete_room(request, room_slug):
 
 @login_required(login_url='login')
 def delete_message(request, pk):
-    message = get_object_or_404(Message, id=pk)
+    message = get_object_or_404(Message, pk=pk)
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == 'POST':
         message.delete()
